+++ UnitTests~/AnimationServices/AnimationIndexTest.cs
@@ -324,5 +324,61 @@ namespace UnitTests.AnimationServices
             Assert.AreEqual(null, newCurve[1].value);
             Assert.AreEqual(m3, newCurve[2].value);
         }
+
+        [Test]
+        public void GetPPtrReferencedObjectsWithBindingTest()
+        {
+            var context = new CloneContext(GenericPlatformAnimatorBindings.Instance);
+            var controller = VirtualAnimatorController.Create(context, "test");
+            var layer = controller.AddLayer(LayerPriority.Default, "test");
+
+            var clip1 = VirtualClip.Create("c1");
+            layer.StateMachine!.AddState("s1", motion: clip1);
+
+            var shader = Shader.Find("Standard");
+            var m1 = new Material(shader) { name = "m1" };
+            var m2 = new Material(shader) { name = "m2" };
+            var ecb = EditorCurveBinding.PPtrCurve("path", typeof(MeshRenderer), "m_Materials.Array.data[0]");
+            clip1.SetObjectCurve(ecb, new ObjectReferenceKeyframe[]
+            {
+                new ObjectReferenceKeyframe { time = 0, value = m1 },
+                new ObjectReferenceKeyframe { time = 1, value = m2 }
+            });
+
+            var index = new AnimationIndex(new[] { controller });
+            var pairs = index.GetPPtrReferencedObjectsWithBinding.ToList();
+            Assert.That(pairs, Has.Exactly(1).Matches<(EditorCurveBinding, Object)>(p => p.Item2 == m1 && p.Item1.Equals(ecb)));
+            Assert.That(pairs, Has.Exactly(1).Matches<(EditorCurveBinding, Object)>(p => p.Item2 == m2 && p.Item1.Equals(ecb)));
+        }
+
+        [Test]
+        public void RewriteObjectCurvesWithBindingTest()
+        {
+            var context = new CloneContext(GenericPlatformAnimatorBindings.Instance);
+            var controller = VirtualAnimatorController.Create(context, "test");
+            var layer = controller.AddLayer(LayerPriority.Default, "test");
+
+            var clip1 = VirtualClip.Create("c1");
+            layer.StateMachine!.AddState("s1", motion: clip1);
+
+            var shader = Shader.Find("Standard");
+            var m1 = new Material(shader) { name = "m1" };
+            var m2 = new Material(shader) { name = "m2" };
+            var m3 = new Material(shader) { name = "m3" };
+            var ecb = EditorCurveBinding.PPtrCurve("path", typeof(MeshRenderer), "m_Materials.Array.data[0]");
+            clip1.SetObjectCurve(ecb, new ObjectReferenceKeyframe[]
+            {
+                new ObjectReferenceKeyframe { time = 0, value = m1 },
+                new ObjectReferenceKeyframe { time = 1, value = m2 }
+            });
+
+            var index = new AnimationIndex(new[] { controller });
+            index.RewriteObjectCurves((binding, obj) => (obj == m2 && binding.Equals(ecb)) ? m3 : obj);
+            var newCurve = clip1.GetObjectCurve(ecb);
+            Assert.AreEqual(2, newCurve.Length);
+            Assert.AreEqual(m1, newCurve[0].value);
+            Assert.AreEqual(m3, newCurve[1].value);
+        }
     }
 }
+