+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added Portable Dynamic Bone Collider component
 - [#644] No alloc overload for `ComputeContext.GetComponentsXXX`
 - [#661] Added `CompatibleWithContext` attribute for plugin passes.
+- [#665] Added `AnimationIndex.GetPPtrReferencedObjectsWithBinding` and a new overload for  
+  `AnimationIndex.RewriteObjectCurves` which provides the curve binding to the rewrite callback.
 
 ### Fixed
 - [#634] Avoid infinite recursion if an avatar is duplicated during play mode activation